Atualização-23/11/2022

+++ trabalhos_ulbra_2022_02/Programação web/docker-lamp-pw-master/www/api/index.php
@@ -4,33 +4,66 @@ $local = $_SERVER['SCRIPT_NAME'];
 $uri = $_SERVER['PHP_SELF']; 
 $rout = str_replace($local, "", $uri);
 $uriSegments = explode("/", $rout);
-// 0-> index.php
-// 1-> clients
-// 2-> id
 
 if(isset($uriSegments[1])){	
 	switch($uriSegments[1]){
+
 		case 'client':
 			require_once("controllers/ClientController.php");
-			$client = new ClientController();
+			$user = new UserController();
+
+			if($user -> isAdmin()){
+				require_once("controllers/ClientController.php");
+				$client = new ClientController();
+				switch($request_method){
+					case 'GET':
+						if(!isset($uriSegments[2]))
+							$client -> listClient();
+						else
+							$client -> consultClient($uriSegments[2]);
+					break;
+					case 'POST':
+						$client ->insertClient();
+					break;
+					case 'PUT':
+						$client ->updateClient($uriSegments[2]);
+					break;
+					case 'DELETE':
+						$client -> deleteClient($uriSegments[2]);
+					break;
+				}
+			}
+		break;	
+
+		case 'user':
+			require_once("controllers/UserController.php");
+			$user = new UserController();
 			switch($request_method){
 				case 'GET':
-					if(!isset($uriSegments[2]))
-						$client -> listClient();
-					else
-						$client -> consultClient($uriSegments[2]);
-				break;
-				case 'POST':
-					$client ->insertClient();
-				break;
-				case 'PUT':
-					$client ->updateClient($uriSegments[2]);
-				break;
-				case 'DELETE':
-					$client -> deleteClient($uriSegments[2]);
+					$user->validateLogin();
 				break;
 			}
 		break;	
+
+		
+        case 'contact':
+            require_once('controllers/ContactController.php');
+            $contactController = new ContactController();
+
+            switch($request_method){
+                case 'POST':
+                    $contactController -> insertContact();
+                break;
+
+                case 'GET':
+                    if(isset($uriSegments[2]) && $uriSegments[2] != '') {
+                        $contactController -> consultContactID($uriSegments[2]);
+                    } else {
+                        $contactController -> consultContact();
+                    }
+                break;
+            }
+        break;
+		
 	}
 }
-?>
